backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
psycopg2-binary==2.9.10
alembic==1.14.0
slowapi==0.1.9
aiosmtplib==3.0.2
pydantic-settings==2.7.1
python-multipart==0.0.20
email-validator==2.2.0
